Update @_specialize and @transpose serialization
…to serialize module selectors.

This commit also changes the serialization format of @_specialize to serialize name information more simply. Instead of using a conditionally-biased argument count and an SPI group count, which could be inconsistent with the number of identifier IDs in the array, we use a kind field and SPI group count and calculate the argument count. This design seems easier to reason about, less likely to become inconsistent, and a little more compact on disk.