Use LLVM runtimes build
LLVM plans to remove the legacy methods for building compiler-rt. This change updates build-script to use LLVM_ENABLE_RUNTIMES to build compiler-rt with the just-built clang instead of the legacy methods.

This does not update the llvm-install-components in
build-presets.ini -- this is done on purpose to stress more the
migration logic, some of the presets will be updated with a subsequent PRs

Addresses rdar://113972453